pr-review-queue: ignore cloud-image-val
We are in this weird situation where we don't own cloud-image-val but
it lives in our org. I would like to lower the amount of review requests
that the bot announces, so let's ignore it. We rarely do reviews there
anyway, so it should be safe.

I also had to reformat the args because the line became too long for the
linter.